[CMPT-5391] complete resolveToken refactor: centralise token logic in auth.go
Move var token, GetToken/SetToken, and resolveToken from get.go into auth.go
so all token-related logic lives in one place. Update post.go, delete.go, and
patch.go to call resolveToken() instead of config.GetAPIKey() directly, making
--skip-auth work consistently for every outbound API verb. Use url.PathEscape
on pipelineID in scope.go with a failing test first to guard against path injection.
